pyproject.toml
[project]
name = "labs"
version = "0.1.0"
description = "rag-service"
requires-python = ">=3.12"
dependencies = [
    "elasticsearch<9",
    "elasticsearch-haystack>=4.1.0",
    "fastapi>=0.122.0",
    "haystack-ai>=2.20.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.2",
    "uvicorn[standard]>=0.38.0",
]
